Edited BillDto and SoldProduct

+++ src/main/java/com/enterprise/application/model/BillDTO.java
@@ -15,5 +15,5 @@ public class BillDTO {
     private String client;
     private String seller;
     private List<SoldProduct> productsSold;
-    private Integer total;
+    private Double total;
 }
